カラムのタイプ修正

+++ database/migrations/2020_03_15_153655_create_check_ins_table.php
@@ -14,9 +14,9 @@ class CreateCheckInsTable extends Migration
     public function up()
     {
         Schema::create('check_ins', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table
-                ->unsignedInteger('user_id')
+                ->unsignedBigInteger('user_id')
                 ->comment('チェックインしたユーザーのID');
             $table->timestamps();
             $table
